Taken out changes into a new file, cloudml_operator_utils.
