+++ tests/contrib/operators/test_cloudml_operator.py
@@ -25,113 +25,13 @@ import httplib2
 import unittest
 
 from airflow import configuration, DAG
-from airflow.contrib.operators import cloudml_operator
 from airflow.contrib.operators.cloudml_operator import CloudMLBatchPredictionOperator
 from airflow.contrib.operators.cloudml_operator import CloudMLTrainingOperator
-from airflow.contrib.operators.cloudml_operator import create_evaluate_ops
 
 from mock import ANY
 from mock import patch
 
 DEFAULT_DATE = datetime.datetime(2017, 6, 6)
-
-
-class CreateEvaluateOpsTest(unittest.TestCase):
-
-    INPUT_MISSING_ORIGIN = {
-        'dataFormat': 'TEXT',
-        'inputPaths': ['gs://legal-bucket/fake-input-path/*'],
-        'outputPath': 'gs://legal-bucket/fake-output-path',
-        'region': 'us-east1',
-    }
-    SUCCESS_MESSAGE_MISSING_INPUT = {
-        'jobId': 'eval_test_prediction',
-        'predictionOutput': {
-            'outputPath': 'gs://fake-output-path',
-            'predictionCount': 5000,
-            'errorCount': 0,
-            'nodeHours': 2.78
-        },
-        'state': 'SUCCEEDED'
-    }
-
-    def setUp(self):
-        super(CreateEvaluateOpsTest, self).setUp()
-        configuration.load_test_config()
-        self.dag = DAG(
-            'test_dag',
-            default_args={
-                'owner': 'airflow',
-                'start_date': DEFAULT_DATE,
-                'end_date': DEFAULT_DATE,
-            },
-            schedule_interval='@daily')
-
-    def testSuccessfulRun(self):
-        input_with_model = self.INPUT_MISSING_ORIGIN.copy()
-        input_with_model['modelName'] = (
-            'projects/test-project/models/test_model')
-        metric_fn = lambda x: (0.1,)
-        metric_fn_encoded = cloudml_operator.base64.b64encode(
-            cloudml_operator.dill.dumps(metric_fn, recurse=True))
-
-        pred, summary, validate = create_evaluate_ops(
-            task_prefix='eval-test',
-            project_id='test-project',
-            job_id='eval-test-prediction',
-            region=input_with_model['region'],
-            data_format=input_with_model['dataFormat'],
-            input_paths=input_with_model['inputPaths'],
-            prediction_path=input_with_model['outputPath'],
-            model_name=input_with_model['modelName'].split('/')[-1],
-            metric_fn_and_keys=(metric_fn, ['err']),
-            validate_fn=(lambda x: 'err=%.1f' % x['err']),
-            dataflow_options=None,
-            dag=self.dag)
-
-        with patch('airflow.contrib.operators.cloudml_operator.'
-                   'CloudMLHook') as mock_cloudml_hook:
-
-            success_message = self.SUCCESS_MESSAGE_MISSING_INPUT.copy()
-            success_message['predictionInput'] = input_with_model
-            hook_instance = mock_cloudml_hook.return_value
-            hook_instance.create_job.return_value = success_message
-            pred.execute(None)
-            mock_cloudml_hook.assert_called_with('google_cloud_default', None)
-            hook_instance.create_job.assert_called_with(
-                'test-project',
-                {
-                    'jobId': 'eval_test_prediction',
-                    'predictionInput': input_with_model
-                })
-
-        with patch('airflow.contrib.operators.dataflow_operator.'
-                   'DataFlowHook') as mock_dataflow_hook:
-
-            hook_instance = mock_dataflow_hook.return_value
-            hook_instance.start_python_dataflow.return_value = None
-            summary.execute(None)
-            mock_dataflow_hook.assert_called_with(
-                gcp_conn_id='google_cloud_default', delegate_to=None)
-            hook_instance.start_python_dataflow.assert_called_with(
-                'eval-test-summary',
-                {
-                    'prediction_path': 'gs://legal-bucket/fake-output-path',
-                    'metric_keys': 'err',
-                    'metric_fn_encoded': metric_fn_encoded,
-                },
-                'airflow.contrib.operators.cloudml_prediction_summary',
-                ['-m'])
-
-        with patch('airflow.contrib.operators.cloudml_operator.'
-                   'GoogleCloudStorageHook') as mock_gcs_hook:
-
-            hook_instance = mock_gcs_hook.return_value
-            hook_instance.download.return_value = '{"err": 0.9, "count": 9}'
-            result = validate.execute({})
-            hook_instance.download.assert_called_with(
-                'legal-bucket', 'fake-output-path/prediction.summary.json')
-            self.assertEqual('err=0.9', result)
 
 
 class CloudMLBatchPredictionOperatorTest(unittest.TestCase):
